refactor: UI 컴포넌트를 개별 파일로 분리
각 UI 컴포넌트 함수를 개별 파일로 분리하여 모듈화 개선:

**생성된 개별 컴포넌트 파일:**
- createProductDropdownSelect.js - 상품 선택 드롭다운
- createCartAddButton.js - 장바구니 추가 버튼
- createStockWarningMessage.js - 재고 경고 메시지
- createProductSelectionPanel.js - 상품 선택 패널
- createCartItemsContainer.js - 장바구니 아이템 컨테이너
- createShoppingAreaColumn.js - 쇼핑 영역 컬럼
- createOrderSummaryColumn.js - 주문 요약 컬럼
- createHelpModalToggleButton.js - 도움말 모달 토글 버튼
- createHelpModalBackdrop.js - 도움말 모달 백드롭
- createHelpContentPanel.js - 도움말 컨텐츠 패널
- createMainLayoutGrid.js - 메인 레이아웃 그리드
- createShoppingCartItemElement.js - 쇼핑카트 아이템 요소

**개선사항:**
- 컴포넌트별 단일 책임 원칙 적용
- 파일명으로 컴포넌트 역할 명확화
- 코드 탐색성 및 유지보수성 향상
- 87개 테스트 모두 통과 확인
