+++ src/basic/components/createCartItemsContainer.js
@@ -1,0 +1,5 @@
+export function createCartItemsContainer() {
+	const div = document.createElement("div");
+	div.id = "cart-items";
+	return div;
+}
