+++ src/basic/components/createHelpModalBackdrop.js
@@ -1,0 +1,11 @@
+export function createHelpModalBackdrop(manualColumn) {
+	const manualOverlay = document.createElement("div");
+	manualOverlay.className = "fixed inset-0 bg-black/50 z-40 hidden transition-opacity duration-300";
+	manualOverlay.onclick = function (e) {
+		if (e.target === manualOverlay) {
+			manualOverlay.classList.add("hidden");
+			manualColumn.classList.add("translate-x-full");
+		}
+	};
+	return manualOverlay;
+}
