+++ src/basic/components/createStockWarningMessage.js
@@ -1,0 +1,6 @@
+export function createStockWarningMessage() {
+	const div = document.createElement("div");
+	div.id = "stock-status";
+	div.className = "text-xs text-red-500 mt-3 whitespace-pre-line";
+	return div;
+}
